- Moved GAN logic into separate directory - Added slots to the OB Widget to start and stop the simulation
