+++ .gitignore
@@ -1,6 +1,6 @@
-cpp codebase/build/
-cpp codebase/buildXCode/
-cpp codebase/buildXcode/
+CXX/build/
+CXX/buildXCode/
+CXX/buildXcode/
 .DS_Store
 DataRetreiver/__pycache__/
 prime_orderbook.db
