+++ CMakeLists.txt
@@ -39,7 +39,7 @@ find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Widgets)
 include_directories(${httplib_SOURCE_DIR})
 include_directories(${torch_SOURCE_DIR})
 include_directories(OrderBook)
-include_directories("cpp codebase/Logic/API")
+include_directories("CXX/Logic/API")
 include_directories(DBHandler)
 include_directories(OrderBookWidget)
 
@@ -51,10 +51,10 @@ message("Going into OrderbookTests subdirectory...")
 #add_subdirectory(OrderBookTests)
 add_subdirectory(OrderBookWidget)
 message("Going into API subdirectory...")
-add_subdirectory("cpp codebase/Logic/API")
+add_subdirectory("CXX/Logic/API")
 
 set(SOURCE_FILES main.cpp)
 add_executable(OrderBookDemo ${SOURCE_FILES})
 
-target_link_libraries(OrderBookDemo PRIVATE OrderBook "cpp codebase/Logic/API" DBHandler OrderBookWidget Qt${QT_VERSION_MAJOR}::Widgets SQLite::SQLite3 httplib::httplib "${TORCH_LIBRARIES}")
+target_link_libraries(OrderBookDemo PRIVATE OrderBook "CXX/Logic/API" DBHandler OrderBookWidget Qt${QT_VERSION_MAJOR}::Widgets SQLite::SQLite3 httplib::httplib "${TORCH_LIBRARIES}")
 set_property(TARGET OrderBookDemo PROPERTY CXX_STANDARD 17)
