+++ CXX/CMakeLists.txt
@@ -1,12 +1,10 @@
-cmake_minimum_required(VERSION 3.14)
-set(CMAKE_CUDA_COMPILER /usr/local/cuda/bin/nvcc)
+cmake_minimum_required(VERSION 3.18)
 project(OrderBookDemo)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(CMAKE_AUTOUIC ON)
 set(CMAKE_AUTOMOC ON)
 set(CMAKE_AUTORCC ON)
-set(CAFFE2_USE_CUDNN 1)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -28,7 +26,6 @@ FetchContent_Declare(
         # URL https://download.pytorch.org/libtorch/cu118/libtorch-cxx11-abi-shared-with-deps-2.2.1%2Bcu118.zip
         URL https://download.pytorch.org/libtorch/cpu/libtorch-cxx11-abi-shared-with-deps-2.2.1%2Bcpu.zip
 )
-
 FetchContent_MakeAvailable(json)
 FetchContent_MakeAvailable(httplib)
 FetchContent_GetProperties(libtorch)
@@ -37,7 +34,7 @@ if(NOT libtorch_POPULATED)
     FetchContent_Populate(libtorch)
     set(Torch_DIR ${libtorch_SOURCE_DIR}/share/cmake/Torch)
 endif()
-
+message(STATUS "Torch_DIR is set to ${Torch_DIR}")
 find_package(Torch REQUIRED)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${TORCH_CXX_FLAGS}")
 
@@ -49,6 +46,7 @@ include_directories(${httplib_SOURCE_DIR})
 include_directories(${torch_SOURCE_DIR})
 include_directories("${CMAKE_SOURCE_DIR}/Logic/OrderBook")
 include_directories("${CMAKE_SOURCE_DIR}/Logic/API")
+include_directories("${CMAKE_SOURCE_DIR}/Logic/GAN")
 include_directories("${CMAKE_SOURCE_DIR}/DummyInput/DBHandler")
 include_directories("${CMAKE_SOURCE_DIR}/UI/OrderBookWidget")
 include_directories("${CMAKE_SOURCE_DIR}/UI/CandlestickWidget")
@@ -59,6 +57,7 @@ include_directories(${torch_INCLUDE_DIRS})
 message("Going into subdirectories...")
 add_subdirectory(Logic/OrderBook)
 add_subdirectory(Logic/API)
+add_subdirectory(Logic/GAN)
 add_subdirectory(DummyInput/DBHandler)
 add_subdirectory(Logic/OrderBookTests)
 
@@ -72,6 +71,5 @@ message("added subdirectories")
 set(SOURCE_FILES main.cpp)
 add_executable(OrderBookDemo ${SOURCE_FILES})
 
-#target_link_libraries(OrderBookDemo PRIVATE OrderBook API DBHandler Qt${QT_VERSION_MAJOR}::Widgets Qt${QT_VERSION_MAJOR}::Charts OrderBookWidget CandlestickWidget CentralWidget HeaderWidget SQLite::SQLite3 "${TORCH_LIBRARIES}")
-target_link_libraries(OrderBookDemo PRIVATE nlohmann_json::nlohmann_json httplib::httplib OrderBook API DBHandler OrderBookWidget CandlestickWidget CentralWidget HeaderWidget Qt${QT_VERSION_MAJOR}::Widgets Qt${QT_VERSION_MAJOR}::Charts SQLite::SQLite3 "${TORCH_LIBRARIES}")
+target_link_libraries(OrderBookDemo PUBLIC ${TORCH_LIBRARIES} nlohmann_json::nlohmann_json httplib::httplib OrderBook API DBHandler GAN OrderBookWidget CandlestickWidget CentralWidget HeaderWidget Qt${QT_VERSION_MAJOR}::Widgets Qt${QT_VERSION_MAJOR}::Charts SQLite::SQLite3)
 set_property(TARGET OrderBookDemo PROPERTY CXX_STANDARD 17)
