+++ CXX/Logic/API/CMakeLists.txt
@@ -3,7 +3,7 @@
 #
 cmake_minimum_required(VERSION 3.12)
 project(API)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 set(HEADER_FILES API.hpp)
 set(SOURCE_FILES API.cpp)
