+++ CXX/UI/CandlestickWidget/CandlestickWidget.cpp
