+++ CXX/UI/CandlestickWidget/CandlestickWidget.hpp
