+++ CXX/UI/HeaderWidget/CMakeLists.txt
@@ -4,4 +4,4 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 add_library(HeaderWidget STATIC HeaderWidget.cpp HeaderWidget.hpp)
-target_link_libraries(HeaderWidget Qt${QT_VERSION_MAJOR}::Widgets)
+target_link_libraries(HeaderWidget GAN Qt${QT_VERSION_MAJOR}::Widgets)
